requirements.txt
astroid==2.0.4
cffi==1.11.5
colorama==0.4.0
isort==4.3.4
lazy-object-proxy==1.3.1
libusb==1.0.22b1
mccabe==0.6.1
pycparser==2.19
pylint==2.1.1
pyusb==1.0.2
sh==1.12.14
six==1.11.0
smbus2==0.2.1
typed-ast==1.1.0
wrapt==1.10.11
gpiozero==1.4.1
RPi.GPIO==0.6.4
spidev==3.2
beanstalkc3==0.4.0
pyyaml==3.13
Flask==1.0.2
Flask-JWT==0.3.2
Flask-API==1.0
Flask-Swagger==0.2.13
Flask-CORS==3.0.6
gphoto2-cffi==0.4.2
fernet==1.0.1
google-api-python-client==1.7.4
httplib2==0.11.3
requests==2.9.1
oauth2client==4.1.3



